chore: remove old otel formatter in lieu of protobuf
