+++ rust/src/adapter/otelstdout.rs
@@ -1,8 +1,10 @@
+use std::vec;
+
 use crate::{Event, TraceEvent};
-use anyhow::{Context, Result};
+use anyhow::Result;
 
 use super::{
-    otel_formatter::{OtelFormatter, ResourceSpan, Span},
+    otel_formatter::{opentelemetry, OtelFormatter},
     Adapter, AdapterHandle,
 };
 
@@ -11,19 +13,13 @@ pub struct OtelStdoutAdapter {}
 
 impl Adapter for OtelStdoutAdapter {
     fn handle_trace_event(&mut self, trace_evt: TraceEvent) -> Result<()> {
-        let mut otf = OtelFormatter::new();
-        let mut rs = ResourceSpan::new();
         let mut spans = vec![];
         let trace_id = trace_evt.telemetry_id.to_hex_16();
         for span in trace_evt.events {
-            self.event_to_spans(&mut spans, span, None, trace_id.clone())?;
+            self.event_to_spans(&mut spans, span, vec![], trace_id.clone())?;
         }
-        rs.add_spans(spans);
-        otf.add_resource_span(rs);
-        println!(
-            "{}",
-            serde_json::to_string(&otf).context("Otel formatter could not create json")?
-        );
+        let otf = OtelFormatter::new(spans, "stdout".into());
+        println!("{:?}", &otf.traces_data);
         Ok(())
     }
 }
@@ -38,17 +34,18 @@ impl OtelStdoutAdapter {
 
     fn event_to_spans(
         &self,
-        spans: &mut Vec<Span>,
+        spans: &mut Vec<opentelemetry::proto::trace::v1::Span>,
         event: Event,
-        parent_id: Option<String>,
+        parent_id: Vec<u8>,
         trace_id: String,
     ) -> Result<()> {
         match event {
             Event::Func(f) => {
                 let name = f.name.clone().unwrap_or("unknown-name".to_string());
 
-                let span = Span::new(trace_id.clone(), parent_id, name, f.start, f.end);
-                let span_id = Some(span.span_id.clone());
+                let span =
+                    OtelFormatter::new_span(trace_id.clone(), parent_id, name, f.start, f.end);
+                let span_id = span.span_id.clone();
                 spans.push(span);
 
                 for e in f.within.iter() {
@@ -56,8 +53,18 @@ impl OtelStdoutAdapter {
                 }
             }
             Event::Alloc(a) => {
-                let mut span = Span::new(trace_id, parent_id, "allocation".to_string(), a.ts, a.ts);
-                span.add_attribute_i64("amount".to_string(), a.amount.into());
+                let mut span = OtelFormatter::new_span(
+                    trace_id,
+                    parent_id,
+                    "allocation".to_string(),
+                    a.ts,
+                    a.ts,
+                );
+                OtelFormatter::add_attribute_i64_to_span(
+                    &mut span,
+                    "amount".to_string(),
+                    a.amount.into(),
+                );
                 spans.push(span);
             }
             _ => {}
